micro/requirements.txt
flask
simplejson
kafka-python
gunicorn